requirements.txt
fastapi
uvicorn
streamlit
openai-whisper
yt-dlp
transformers
torch
scipy
requests
python-dotenv
httpx
pydantic
deno